# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafecitos API WEB",
        "description": "API for Cafecitos",
        "contact": {
            "email": "devcitobo@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-waiters": {
            "post": {
                "tags": [
                    "OrderWaiters"
                ],
                "summary": "Create a new order waiter assignment",
                "operationId": "9ab8935b153d89a1b15a684478c083f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_det_id",
                                    "waiter_id"
                                ],
                                "properties": {
                                    "order_det_id": {
                                        "type": "integer"
                                    },
                                    "waiter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order waiter created"
                    },
                    "409": {
                        "description": "Order is closed"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/order-waiters/{id}": {
            "get": {
                "tags": [
                    "OrderWaiters"
                ],
                "summary": "Get an order waiter assignment by ID",
                "operationId": "7d95144c93e8c51657bd3de8efffe50d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Order waiter not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "OrderWaiters"
                ],
                "summary": "Update an order waiter assignment",
                "operationId": "83fe8fa14ae14c5b0ea1bf394254dcb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "waiter_id"
                                ],
                                "properties": {
                                    "waiter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order waiter updated"
                    },
                    "409": {
                        "description": "Order is closed"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OrderWaiters"
                ],
                "summary": "Delete an order waiter assignment",
                "operationId": "d746edd914b715e44ca17cbc638733dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order waiter deleted"
                    },
                    "404": {
                        "description": "Order waiter not found"
                    },
                    "409": {
                        "description": "Order is closed"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "csrf": {
                "type": "apiKey",
                "description": "Para obtener este token, primero asegúrate de haber iniciado sesión. Luego, ejecuta el endpoint GET /sanctum/csrf-cookie. Finalmente, abre las herramientas de desarrollador de tu navegador, busca la cookie 'XSRF-TOKEN', copia su valor y pégalo aquí. Este token es necesario para todas las peticiones que modifican datos (POST, PUT, PATCH, DELETE).",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}